refactor(projectCardSkeleton): shift projectCardSkeleton under /Skeletons/Project
